move fetchRealDemoUrl to faceit.ts
