+++ src/api/analyzer.ts
@@ -1,5 +1,6 @@
 import { z } from "zod";
 import camelcaseKeys from "camelcase-keys";
+import { fetchRealDemoUrl } from "./faceit";
 
 const analyzerStatusEnum = z.enum([
   "waiting",
@@ -28,16 +29,6 @@ const analyzerGameStatusSchema = z.object({
 });
 export type AnalyzerGameStatus = z.infer<typeof analyzerGameStatusSchema>;
 
-const realDemoUrlSchema = z
-  .object({
-    payload: z.object({
-      download_url: z.string(),
-    }),
-  })
-  .transform((data) => {
-    return camelcaseKeys(data, { deep: true });
-  });
-
 const sendDemoUrlResponseSchema = z
   .object({
     demo_id: z.string(),
@@ -58,43 +49,6 @@ export type AnalyzerMatchStatus = z.infer<typeof analyzerMatchStatusSchema>;
 const BASE_URL = "csanalyzer.gg";
 const COLLECTOR_URL = `https://collector.${BASE_URL}`;
 const ART_URL = `https://art.${BASE_URL}`;
-
-export interface CustomError extends Error {
-  code: number;
-}
-
-const fetchRealDemoUrl = async (demoUrl: string) => {
-  const url = "https://www.faceit.com/api/download/v2/demos/download-url";
-  const payload = { resource_url: demoUrl };
-
-  try {
-    const response = await fetch(url, {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify(payload),
-    });
-
-    if (!response.ok) {
-      const error = new Error(`HTTP error: ${response.status}`) as CustomError;
-      error.code = response.status;
-      throw error;
-    }
-
-    const json = await response.json();
-    const parsed = realDemoUrlSchema.parse(json);
-    return parsed.payload.downloadUrl;
-  } catch (error: unknown) {
-    if (typeof error === "object" && error !== null && "message" in error) {
-      const typedError = error as CustomError;
-      if (typedError.code === undefined) typedError.code = 0;
-      console.error("Error fetching real demo URL:", typedError.message);
-      throw typedError;
-    } else {
-      console.error("Unknown error:", error);
-      throw error;
-    }
-  }
-};
 
 export const fetchAnalyzerGameStatus = async (matchId: string) => {
   const url = `${COLLECTOR_URL}/faceit/matches/${matchId}`;
